Made app ui more modular by relocating tobacco adding logic and ui to fragment and adding it to main activity
