+++ app/src/main/res/layout/fragment_tobacco_settings.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/root_view"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+
+    tools:context=".TobaccoSettingsFragment">
+
+    <include
+        android:id="@+id/grid_dry"
+        layout="@layout/card_grid_dry_properties"
+        android:layout_width="210dp"
+        android:layout_height="250dp"
+        android:layout_marginTop="16dp"
+        android:layout_marginEnd="4dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <include
+        android:id="@+id/grid_wet"
+        layout="@layout/card_grid_wet_properties"
+        android:layout_width="210dp"
+        android:layout_height="225dp"
+        android:layout_marginEnd="4dp"
+        android:layout_marginBottom="16dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent" />
+
+    <include
+        android:id="@+id/leaves_contents"
+        layout="@layout/card_leaves_contents"
+        android:layout_width="159dp"
+        android:layout_height="158dp"
+        android:layout_marginTop="16dp"
+        android:layout_marginEnd="12dp"
+        android:visibility="invisible"
+        app:layout_constraintEnd_toStartOf="@+id/grid_dry"
+        app:layout_constraintHorizontal_bias="0.615"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
