+++ app/src/main/res/values/strings.xml
@@ -14,4 +14,6 @@
     <string name="virginia_concentration">Virginia:</string>
     <string name="save_recipe">Save the recipe</string>
     <string name="big_tare">Big tare:</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
